WIP - Vue progress & added payment calculator

+++ database/migrations/2023_06_10_231740_add_fields_to_listings_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::table('listings', function (Blueprint $table) {
             $table->unsignedBigInteger('beds');
             $table->unsignedBigInteger('baths');
-            $table->unsignedBigInteger('area');
+            $table->unsignedSmallInteger('area');
 
             $table->tinyText('city');
             $table->tinyText('code');
